+++ tests/testthat/test-apropos.R
@@ -1,0 +1,5 @@
+test_that("apropos returns formatted strings", {
+		  maxima.start()
+  expect_match(maxima.apropos("int"), "^\\$\\$[[:space:]|[:print:]]*\\$\\$$")
+})
+